cdh/KMS: add support for aliyun STS token client
This patch will add STS token support for aliyun KMS client. The part of
STS token code is also be reused by ecs_ram_role client. Also does some
refactoring work upon the code.
